New Sending protocoll, new Class. Problem: Client is not receiving any result

+++ A4/VSP4_Mi/src/math_ops/_CalculatorImplBase.java
@@ -2,7 +2,7 @@ package math_ops;
 
 public interface _CalculatorImplBase{
 
-	public abstract double add(double a, double b);
+	public abstract int add(int a, int b);
 	
 	public static _CalculatorImplBase narrowCast(Object rawObjectRef){
 		return new _CalculatorImplBaseStub(rawObjectRef);
